\name{emboss}

\alias{emboss}

\concept{embossing}

\title{Image embossing}

\description{
  This function generates an embossed image.
}

\usage{
emboss(img, x = 2, y = 2)
}

\arguments{
  \item{img}{An image of \code{Image} class.} 
  \item{x, y}{Amount of translation in pixel.}
}

\value{
  Returns an \code{Image} class object.
}

\examples{
  demo(embossDemo)
}

\author{
  Yuichiro Hourai, \email{yhourai@gmail.com}
  
  Takeo Katsuki, \email{takeo.katsuki@gmail.com}
}

